refactor: Move components to different folders
Moved all of the pages you see while logged out to a separate folder
since at least two of them share a common component and I didn't know
where to put it lol.
Also moved the password toggle button to a separate file.
