+++ src/components/pages/loggedOut/createAccountPage/CreateAccount.tsx
@@ -9,7 +9,7 @@ import {
   IconButton,
   InputAdornment,
 } from "@mui/material";
-import BACKEND_URL from "../../../backendUrl";
+import BACKEND_URL from "../../../../backendUrl";
 import { useNavigate } from "react-router-dom";
 import {
   loginButton,
@@ -17,10 +17,11 @@ import {
   loginRegisterTitle,
   loginRegisterContent,
   AuthBoxTitle,
-} from "../../../sxStyles";
-import { setSession } from "../../../auth";
-import { TheContext } from "../../../TheContext";
+} from "../../../../sxStyles";
+import { setSession } from "../../../../auth";
+import { TheContext } from "../../../../TheContext";
 import { VisibilityOff, Visibility } from "@mui/icons-material";
+import PasswordToggle from "../passwordToggleButton/PasswordToggle";
 
 const REQUIRED_PASSWORD_LENGTH = 8;
 
@@ -210,20 +211,11 @@ const CreateAccount: FC = () => {
                     margin="normal"
                     InputProps={{
                       endAdornment: (
-                        <InputAdornment position="end">
-                          <IconButton
-                            aria-label="toggle password visibility"
-                            onClick={() => handleClickShowPassword("first")}
-                            onMouseDown={handleMouseDownPassword}
-                            edge="end"
-                          >
-                            {showPassword.first ? (
-                              <VisibilityOff />
-                            ) : (
-                              <Visibility />
-                            )}
-                          </IconButton>
-                        </InputAdornment>
+                        <PasswordToggle
+                          onMouseDown={handleMouseDownPassword}
+                          onClick={() => handleClickShowPassword("first")}
+                          passwordVisible={showPassword.first}
+                        ></PasswordToggle>
                       ),
                     }}
                     onChange={inputChange}
@@ -236,20 +228,11 @@ const CreateAccount: FC = () => {
                     margin="normal"
                     InputProps={{
                       endAdornment: (
-                        <InputAdornment position="end">
-                          <IconButton
-                            aria-label="toggle password visibility"
-                            onClick={() => handleClickShowPassword("second")}
-                            onMouseDown={handleMouseDownPassword}
-                            edge="end"
-                          >
-                            {showPassword.second ? (
-                              <VisibilityOff />
-                            ) : (
-                              <Visibility />
-                            )}
-                          </IconButton>
-                        </InputAdornment>
+                        <PasswordToggle
+                          onMouseDown={handleMouseDownPassword}
+                          onClick={() => handleClickShowPassword("second")}
+                          passwordVisible={showPassword.second}
+                        ></PasswordToggle>
                       ),
                     }}
                     onChange={inputChange}
